+++ bench/bench-java/src/main/java/com/softwaremill/jox/BufferedBenchmark.java
@@ -1,13 +1,13 @@
-package jox;
+package com.softwaremill.jox;
 
+import com.softwaremill.jox.Channel;
 import org.openjdk.jmh.annotations.*;
 
-import java.util.concurrent.Exchanger;
-import java.util.concurrent.SynchronousQueue;
+import java.util.concurrent.ArrayBlockingQueue;
 import java.util.concurrent.TimeUnit;
 
 /**
- * Rendezvous tests for {@link SynchronousQueue}, {@link Exchanger} and {@link Channel}.
+ * Buffered tests for {@link ArrayBlockingQueue} and {@link Channel}.
  */
 @Warmup(iterations = 3, time = 5000, timeUnit = TimeUnit.MILLISECONDS)
 @Measurement(iterations = 10, time = 5000, timeUnit = TimeUnit.MILLISECONDS)
@@ -18,44 +18,36 @@ import java.util.concurrent.TimeUnit;
 @BenchmarkMode(Mode.AverageTime)
 @OutputTimeUnit(TimeUnit.NANOSECONDS)
 @State(Scope.Group)
-public class RendezvousBenchmark {
-    private SynchronousQueue<Integer> queue = new SynchronousQueue<>();
+public class BufferedBenchmark {
+    @Param({"1", "10", "100"})
+    public int capacity;
+
+    private ArrayBlockingQueue<Integer> queue;
+    private Channel<Integer> channel;
+
+    @Setup
+    public void create() {
+        queue = new ArrayBlockingQueue<>(capacity);
+        channel = new Channel<>(capacity);
+    }
+
+    //
 
     @Benchmark
-    @Group("synchronous_queue")
+    @Group("array_blocking_queue")
     @GroupThreads(1)
-    public void putToSynchronousQueue() throws InterruptedException {
+    public void putToArrayBlockingQueue() throws InterruptedException {
         queue.put(63);
     }
 
     @Benchmark
-    @Group("synchronous_queue")
+    @Group("array_blocking_queue")
     @GroupThreads(1)
-    public void takeFromSynchronousQueue() throws InterruptedException {
+    public void takeFromArrayBlockingQueue() throws InterruptedException {
         queue.take();
     }
 
     //
-
-    private Exchanger<Integer> exchanger = new Exchanger<>();
-
-    @Benchmark
-    @Group("exchanger")
-    @GroupThreads(1)
-    public void exchange1() throws InterruptedException {
-        exchanger.exchange(63);
-    }
-
-    @Benchmark
-    @Group("exchanger")
-    @GroupThreads(1)
-    public void exchange2() throws InterruptedException {
-        exchanger.exchange(64);
-    }
-
-    //
-
-    private Channel<Integer> channel = new Channel<>();
 
     @Benchmark
     @Group("channel")
