+++ bench/bench-kotlin/src/com/softwaremill/jox/RendezvousKotlinBenchmark.kt
@@ -1,4 +1,4 @@
-package jox
+package com.softwaremill.jox
 
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.channels.Channel
@@ -14,16 +14,12 @@ import java.util.concurrent.TimeUnit
 @Fork(value = 3)
 @BenchmarkMode(Mode.AverageTime)
 @OutputTimeUnit(TimeUnit.NANOSECONDS)
-@State(Scope.Benchmark)
-open class BufferedKotlinBenchmark {
-    @Param("1", "10", "100")
-    var capacity: Int = 0
-
+open class RendezvousKotlinBenchmark {
     @Benchmark
     @OperationsPerInvocation(OPERATIONS_PER_INVOCATION)
     fun sendReceiveUsingDefaultDispatcher() {
         runBlocking {
-            val channel = Channel<Long>(capacity)
+            val channel = Channel<Long>(0)
             launch(Dispatchers.Default) {
                 for (x in 1..OPERATIONS_PER_INVOCATION) channel.send(63)
                 channel.close()
