+++ bench/bench-kotlin/src/com/softwaremill/jox/constants.kt
@@ -1,3 +1,3 @@
-package jox
+package com.softwaremill.jox
 
 const val OPERATIONS_PER_INVOCATION = 1_000_000
