+++ core/src/test/java/com/softwaremill/jox/ChannelBufferedTest.java
@@ -1,12 +1,12 @@
-package jox;
+package com.softwaremill.jox;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.Timeout;
 
 import java.util.concurrent.ExecutionException;
 
-import static jox.TestUtil.forkVoid;
-import static jox.TestUtil.scoped;
+import static com.softwaremill.jox.TestUtil.forkVoid;
+import static com.softwaremill.jox.TestUtil.scoped;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertInstanceOf;
