+++ core/src/test/java/com/softwaremill/jox/ChannelStressTest.java
@@ -1,9 +1,9 @@
-package jox;
+package com.softwaremill.jox;
 
 import java.util.*;
 import java.util.concurrent.*;
 
-import static jox.TestUtil.*;
+import static com.softwaremill.jox.TestUtil.*;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
