+++ core/src/test/java/com/softwaremill/jox/SegmentTest.java
@@ -1,8 +1,8 @@
-package jox;
+package com.softwaremill.jox;
 
 import org.junit.jupiter.api.Test;
 
-import static jox.Segment.SEGMENT_SIZE;
+import static com.softwaremill.jox.Segment.SEGMENT_SIZE;
 import static org.junit.jupiter.api.Assertions.*;
 
 public class SegmentTest {
